Module identity_credential::credential

The core types used to create Verifiable Credentials

Modules

builder

credential

evidence

issuer

policy

refresh

schema

status

subject

verifiable

Structs

Credential
A Credential represents a set of claims describing an entity.

CredentialBuilder
A CredentialBuilder is used to create a customized Credential.

Evidence
Information used to increase confidence in the claims of a Credential

Policy
Information used to express obligations, prohibitions, and permissions about a Credential or Presentation.

Refresh
Information used to refresh or assert the status of a Credential.

Schema
Information used to validate the structure of a Credential.

Status
Information used to determine the current status of a Credential.

Subject
An entity who is the target of a set of claims.

VerifiableCredential
A VerifiableCredential represents a Credential with an associated digital proof.

Enums

Issuer
An identifier representing the issuer of a Credential.